requirements.txt
django>=4.0.0,<4.2.4
psycopg>=3.1.4,<3.2.0
djangorestframework>=3.14.0,<3.15.0
drf-spectacular>=0.26.5,<0.27.0

